move constellation "open" state to cache
Constellation open/closed state has been tracked in the main settings object,
but this meant that every change to open constellations triggered a diff if
you track the file in git. It also created a logical contradiction between
separate systems if you didn't want the same thing open on each. This state is
now tracked in a standalone cache file.